election def xml parsing and data loading

+++ app/services/data_loader.rb
@@ -10,6 +10,7 @@ class DataLoader
     load_election
     load_districts
     load_precincts
+    load_candidates
   end
 
   private
@@ -17,17 +18,17 @@ class DataLoader
   def load_election
     election_el = @doc.css("vip_object > election").first
 
-    uid = election_el['id']
-    state_uid = election_el.css("> state_id").first.content
-    date = election_el.css("> date").first.content
-    type = election_el.css("> election_type").first.content
-    statewide = election_el.css("> statewide").first.content == "YES"
+    uid       = election_el['id']
+    state_uid = dequote(election_el.css("> state_id").first.content)
+    date      = dequote(election_el.css("> date").first.content)
+    type      = dequote(election_el.css("> election_type").first.content)
+    statewide = dequote(election_el.css("> statewide").first.content).upcase == "YES"
 
     state = State.find_by_uid!(state_uid)
     state.elections.create_with({
-      held_on: date,
-      election_type: type,
-      statewide: statewide
+      held_on:        date,
+      election_type:  type,
+      statewide:      statewide
     }).find_or_create_by(uid: uid)
   end
 
@@ -39,40 +40,44 @@ class DataLoader
     end
   end
 
-  def load_precincts
+  def for_each_locality
     for_each_state do |state_el, state|
       state_el.css("locality").each do |locality_el|
-        uid = locality_el['id']
-        name = locality_el.css('> name').first.content.titleize
-        type = locality_el.css('> type').first.content
+        uid      = locality_el['id']
+        name     = dequote(locality_el.css('> name').first.content).titleize
+        type     = dequote(locality_el.css('> type').first.content)
         locality = state.localities.create_with(name: name, locality_type: type).find_or_create_by(uid: uid)
 
-        # continue loading precincts
-        locality.precincts.destroy_all
-        locality_el.css('precinct').each do |precinct_el|
-          uid = precinct_el['id']
-          name = dequote(precinct_el.css('> name').first.content)
-          precinct = locality.precincts.create_with(name: name).find_or_create_by(uid: uid)
+        yield locality_el, locality
+      end
+    end
+  end
 
-          precinct_el.css('> electoral_district_id').each do |electoral_district_id_el|
-            uid = electoral_district_id_el.content
-            precinct.districts << @districts[uid]
-          end
+  def load_precincts
+    return if @doc.css('vip_object > state > locality > precinct').size == 0
 
-          create_polling_location(precinct_el, precinct)
+    for_each_locality do |locality_el, locality|
+      # continue loading precincts
+      locality.precincts.destroy_all
+      locality_el.css('precinct').each do |precinct_el|
+        uid      = precinct_el['id']
+        name     = dequote(precinct_el.css('> name').first.content)
+        precinct = locality.precincts.create_with(name: name).find_or_create_by(uid: uid)
+
+        precinct_el.css('> electoral_district_id').each do |electoral_district_id_el|
+          uid = electoral_district_id_el.content
+          precinct.districts << @districts[uid]
         end
+
+        create_polling_location(precinct_el, precinct)
       end
     end
   end
 
   def load_districts
-    @doc.css('electoral_district').each do |district_el|
-      uid = district_el['id']
-      name = district_el.css('> name').first.content
-      type = district_el.css('> type').first.content
-      district = District.create_with(name: name, district_type: type).find_or_create_by(uid: uid)
-
-      @districts[uid] = district
+    @doc.css('vip_object > electoral_district').each do |district_el|
+      district = find_or_create_district(district_el)
+      @districts[district.uid] = district
     end
   end
 
@@ -81,17 +86,52 @@ class DataLoader
     address_el = polling_location_el.css('> address').first
 
     precinct.create_polling_location({
-      name:  polling_location_el.css('> name').first.content,
-      line1: address_el.css('> line1').first.content,
-      line2: address_el.css('> line2').first.try(:content),
-      city:  address_el.css('> city').first.content,
-      state: address_el.css('> state').first.content,
-      zip:   address_el.css('> zip').first.content
+      name:  dequote(polling_location_el.css('> name').first.content),
+      line1: dequote(address_el.css('> line1').first.content),
+      line2: dequote(address_el.css('> line2').first.try(:content)),
+      city:  dequote(address_el.css('> city').first.content),
+      state: dequote(address_el.css('> state').first.content),
+      zip:   dequote(address_el.css('> zip').first.content)
     })
   end
 
+  def load_candidates
+    return if @doc.css('vip_object > state > locality > contest').size == 0
+
+    for_each_contest do |contest_el, contest|
+      contest_el.css("> candidate").each do |candidate_el|
+        uid        = candidate_el['id']
+        name       = dequote(candidate_el.css('> name').first.content)
+        party      = dequote(candidate_el.css('> party').first.content)
+        sort_order = dequote(candidate_el.css('> sort_order').first.content)
+
+        contest.candidates.create_with(name: name, party: party, sort_order: sort_order).find_or_create_by(uid: uid)
+      end
+    end
+  end
+
+  def for_each_contest(&block)
+    for_each_locality do |locality_el, locality|
+      locality_el.css("> contest").each do |contest_el|
+        uid        = contest_el['id']
+        office     = dequote(contest_el.css("> office").first.content)
+        sort_order = dequote(contest_el.css("> sort_order").first.content)
+        district   = find_or_create_district(contest_el.css("> electoral_district").first)
+        contest    = locality.contests.create_with(office: office, sort_order: sort_order, district: district).find_or_create_by(uid: uid)
+
+        block.call(contest_el, contest)
+      end
+    end
+  end
+
+  def find_or_create_district(district_el)
+    name = dequote(district_el.css("> name").first.content)
+    type = dequote(district_el.css("> type").first.content)
+    District.create_with(name: name, district_type: type).find_or_create_by(uid: district_el['id'])
+  end
+
   def dequote(v)
-    v.gsub(/(^["']|["']$)/, '')
+    v.blank? ? v : v.gsub(/(^["']|["']$)/, '')
   end
 
 end
